Create synchronization history service for client server.
